requirements.txt
fastapi[all]
uvicorn
sentence-transformers
cachetools
numpy
python-multipart
python-dotenv